Fixed usage of incorrect type in the 'usePluginComponents' hook.
