+++ packages/grafana-runtime/src/services/index.ts
@@ -21,9 +21,6 @@ export {
   type GetPluginExtensionsResult,
   type UsePluginExtensions,
   type UsePluginExtensionsResult,
-  type UsePluginComponentResult,
-  type UsePluginFunctionsOptions,
-  type UsePluginFunctionsResult,
 } from './pluginExtensions/getPluginExtensions';
 export {
   setPluginExtensionsHook,
@@ -32,10 +29,29 @@ export {
   usePluginComponentExtensions,
 } from './pluginExtensions/usePluginExtensions';
 
-export { setPluginComponentHook, usePluginComponent } from './pluginExtensions/usePluginComponent';
-export { setPluginComponentsHook, usePluginComponents } from './pluginExtensions/usePluginComponents';
-export { setPluginLinksHook, usePluginLinks } from './pluginExtensions/usePluginLinks';
-export { setPluginFunctionsHook, usePluginFunctions } from './pluginExtensions/usePluginFunctions';
+export {
+  setPluginComponentHook,
+  usePluginComponent,
+  type UsePluginComponentResult,
+} from './pluginExtensions/usePluginComponent';
+export {
+  setPluginComponentsHook,
+  usePluginComponents,
+  type UsePluginComponentsResult,
+  type UsePluginComponentsOptions,
+} from './pluginExtensions/usePluginComponents';
+export {
+  setPluginLinksHook,
+  usePluginLinks,
+  type UsePluginLinksOptions,
+  type UsePluginLinksResult,
+} from './pluginExtensions/usePluginLinks';
+export {
+  setPluginFunctionsHook,
+  usePluginFunctions,
+  type UsePluginFunctionsOptions,
+  type UsePluginFunctionsResult,
+} from './pluginExtensions/usePluginFunctions';
 
 export { isPluginExtensionLink, isPluginExtensionComponent } from './pluginExtensions/utils';
 export { setCurrentUser } from './user';
