+++ packages/grafana-runtime/src/services/pluginExtensions/usePluginComponents.ts
@@ -1,13 +1,23 @@
-import { GetPluginExtensionsOptions, UsePluginComponentsResult } from './getPluginExtensions';
+import { PluginExtensionComponentMeta } from '@grafana/data';
+
+export type UsePluginComponentsOptions = {
+  extensionPointId: string;
+  limitPerPlugin?: number;
+};
+
+export type UsePluginComponentsResult<Props = {}> = {
+  components: Array<React.ComponentType<Props> & { meta: PluginExtensionComponentMeta }>;
+  isLoading: boolean;
+};
 
 export type UsePluginComponents<Props extends object = {}> = (
-  options: GetPluginExtensionsOptions
+  options: UsePluginComponentsOptions
 ) => UsePluginComponentsResult<Props>;
 
 let singleton: UsePluginComponents | undefined;
 
 export function setPluginComponentsHook(hook: UsePluginComponents): void {
-  // We allow overriding the registry in tests
+  // We allow overriding the hook in tests
   if (singleton && process.env.NODE_ENV !== 'test') {
     throw new Error('setPluginComponentsHook() function should only be called once, when Grafana is starting.');
   }
@@ -15,7 +25,7 @@ export function setPluginComponentsHook(hook: UsePluginComponents): void {
 }
 
 export function usePluginComponents<Props extends object = {}>(
-  options: GetPluginExtensionsOptions
+  options: UsePluginComponentsOptions
 ): UsePluginComponentsResult<Props> {
   if (!singleton) {
     throw new Error('setPluginComponentsHook(options) can only be used after the Grafana instance has started.');
